add force-decoding-eos token logit processor
