+++ llm/predictor.py
@@ -74,13 +74,11 @@ class PredictorArgument:
     inference_model: bool = field(default=False, metadata={"help": "whether use InferenceModel to do generation"})
     batch_size: int = field(default=1, metadata={"help": "The batch size of data."})
     max_batch_size: int = field(default=None, metadata={"help": "The max batch size of data during serving."})
-    benchmark: bool = (
-        field(
-            default=False,
-            metadata={
-                "help": "If benchmark set as `True`, we will force model decode to max_length, which is helpful to compute throughput. "
-            },
-        ),
+    benchmark: bool = field(
+        default=False,
+        metadata={
+            "help": "If benchmark set as `True`, we will force model decode to max_length, which is helpful to compute throughput. "
+        },
     )
