+++ paddlenlp/transformers/llama/configuration.py
@@ -25,6 +25,7 @@ __all__ = [
 LLAMA_PRETRAINED_INIT_CONFIGURATION = {
     # Hypothetical model weights (tiny-random-llama & micro-random-llama) for test only
     "__internal_testing__/micro-random-llama": {
+        "architectures": ["LlamaForCausalLM"],
         "hidden_size": 64,
         "initializer_range": 0.02,
         "intermediate_size": 1000,
@@ -42,6 +43,7 @@ LLAMA_PRETRAINED_INIT_CONFIGURATION = {
         "use_flash_attention": False,
     },
     "__internal_testing__/tiny-random-llama": {
+        "architectures": ["LlamaForCausalLM"],
         "hidden_size": 768,
         "initializer_range": 0.02,
         "intermediate_size": 11008,
